+++ server/src/modules/session/service.ts
@@ -19,6 +19,8 @@ import { OrganizationRepository } from '@modules/organizations/repository';
 import { OrganizationUsersRepository } from '@modules/organization-users/repository';
 import { fullName, generateOrgInviteURL, isSuperAdmin } from '@helpers/utils.helper';
 import { decamelizeKeys } from 'humps';
+import { RequestContext } from '@modules/request-context/service';
+import { AUDIT_LOGS_REQUEST_CONTEXT_KEY } from '@modules/app/constants';
 
 @Injectable()
 export class SessionService {
@@ -34,7 +36,17 @@ export class SessionService {
     response.clearCookie('tj_auth_token');
     await dbTransactionWrap(async (manager: EntityManager) => {
       await manager.delete(UserSessions, { id: sessionId, userId });
-      await this.sessionUtilService.setAuditLogForUser(userId, manager);
+      const user = await manager.findOneOrFail(User, {
+        where: { id: userId },
+      });
+
+      const auditLogData = {
+        userId: user.id,
+        organizationId: user.defaultOrganizationId,
+        resourceId: user.id,
+        resourceName: user.email,
+      };
+      RequestContext.setLocals(AUDIT_LOGS_REQUEST_CONTEXT_KEY, auditLogData);
     });
   }
